[git-p4: depot-paths = "//SW_Components/Mainline/services/DAL/Src/": change = 417302]

+++ plugins/bhplugin1/jhi_plugin.cpp
@@ -295,7 +295,7 @@ end:
 				versionQuery = convertAppProperty_Version(&output); // convert to unsigned int like in TL.
 			}
 
-			outputLength = strlen(output);
+			outputLength = (int)strlen(output);
 
 			if (*outputBufferLength < outputLength)
 			{
